# swagger.yml
swagger: '2.0'
info:
  title: forum
  description: |
    Тестовое задание для реализации проекта "Форумы" на курсе по базам данных в
    Технопарке VK (https://park.vk.company).
  version: "0.1.0"
schemes:
  - http
  - https
basePath: /api
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /thread/{slug_or_id}/create:
    post:
      summary: Создание новых постов
      tags:
        - thread
      description: |
        Добавление новых постов в ветку обсуждения на форум.

        Все посты, созданные в рамках одного вызова данного метода должны иметь одинаковую дату создания (Post.Created).
      operationId: postsCreate
      parameters:
        - name: slug_or_id
          in: path
          description: Идентификатор ветки обсуждения.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: posts
          in: body
          description: Список создаваемых постов.
          required: true
          schema:
            $ref: '#/components/schemas/Posts'
      responses:
        201:
          description: |
            Посты успешно созданы.
            Возвращает данные созданных постов в том же порядке, в котором их передали на вход метода.
          schema:
            $ref: '#/components/schemas/Posts'
        404:
          description: |
            Ветка обсуждения отсутствует в базе данных.
          schema:
            $ref: '#/components/schemas/Error'
        409:
          description: |
            Хотя бы один родительский пост отсутсвует в текущей ветке обсуждения.
          schema:
            $ref: '#/components/schemas/Error'
  /thread/{slug_or_id}/details:
    get:
      summary: Получение информации о ветке обсуждения
      tags:
        - thread
      description: |
        Получение информации о ветке обсуждения по его имени.
      consumes: [ ]
      operationId: threadGetOne
      parameters:
        - name: slug_or_id
          in: path
          description: Идентификатор ветки обсуждения.
          required: true
          type: string
          schema:
            $ref: '#/components/schemas/Slug'
      responses:
        200:
          description: |
            Информация о ветке обсуждения.
          schema:
            $ref: '#/components/schemas/Thread'
        404:
          description: |
            Ветка обсуждения отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
    post:
      summary: Обновление ветки
      tags:
        - thread
      description: |
        Обновление ветки обсуждения на форуме.
      operationId: threadUpdate
      parameters:
        - name: slug_or_id
          in: path
          description: Идентификатор ветки обсуждения.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: thread
          in: body
          description: Данные ветки обсуждения.
          required: true
          schema:
            $ref: '#/components/schemas/ThreadUpdate'
      responses:
        200:
          description: |
            Информация о ветке обсуждения.
          schema:
            $ref: '#/components/schemas/Thread'
        404:
          description: |
            Ветка обсуждения отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
  /thread/{slug_or_id}/posts:
    get:
      summary: Сообщения данной ветви обсуждения
      tags:
        - thread
      description: |
        Получение списка сообщений в данной ветке форуме.

        Сообщения выводятся отсортированные по дате создания.
      consumes: [ ]
      operationId: threadGetPosts
      parameters:
        - name: slug_or_id
          in: path
          description: Идентификатор ветки обсуждения.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: limit
          in: query
          type: number
          
          default: 100
          minimum: 1
          maximum: 10000
          description: Максимальное кол-во возвращаемых записей.
          schema:
            $ref: '#/components/schemas/Limit'
        - name: since
          in: query
          type: number
          
          description: |
            Идентификатор поста, после которого будут выводиться записи
            (пост с данным идентификатором в результат не попадает).
          schema:
            $ref: '#/components/schemas/Since'
        - name: sort
          in: query
          type: string
          description: |
            Вид сортировки:

             * flat - по дате, комментарии выводятся простым списком в порядке создания;
             * tree - древовидный, комментарии выводятся отсортированные в дереве
               по N штук;
             * parent_tree - древовидные с пагинацией по родительским (parent_tree),
               на странице N родительских комментов и все комментарии прикрепленные
               к ним, в древвидном отображение.

            Подробности: https://park.mail.ru/blog/topic/view/1191/
          default: flat
          enum:
            - flat
            - tree
            - parent_tree
          schema:
            $ref: '#/components/schemas/Sort'
        - name: desc
          in: query
          type: boolean
          description: |
            Флаг сортировки по убыванию.
          schema:
            $ref: '#/components/schemas/Desc'
      responses:
        200:
          description: |
            Информация о сообщениях форума.
          schema:
            $ref: '#/components/schemas/Posts'
        404:
          description: |
            Ветка обсуждения отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
  /thread/{slug_or_id}/vote:
    post:
      summary: Проголосовать за ветвь обсуждения
      tags:
        - thread
      description: |
        Изменение голоса за ветвь обсуждения.

        Один пользователь учитывается только один раз и может изменить своё
        мнение.
      operationId: threadVote
      parameters:
        - name: slug_or_id
          in: path
          description: Идентификатор ветки обсуждения.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: vote
          in: body
          description: Информация о голосовании пользователя.
          required: true
          schema:
            $ref: '#/components/schemas/Vote'
      responses:
        200:
          description: |
            Информация о ветке обсуждения.
          schema:
            $ref: '#/components/schemas/Thread'
        404:
          description: |
            Ветка обсуждения отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
  /user/{nickname}/create:
    post:
      summary: Создание нового пользователя
      tags:
        - user
      description: |
        Создание нового пользователя в базе данных.
      operationId: userCreate
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
          schema:
            $ref: '#/components/schemas/Nickname'
        - name: profile
          in: body
          description: Данные пользовательского профиля.
          required: true
          schema:
            $ref: '#/components/schemas/User'
      responses:
        201:
          description: |
            Пользователь успешно создан.
            Возвращает данные созданного пользователя.
          schema:
            $ref: '#/components/schemas/User'
        409:
          description: |
            Пользователь уже присутсвует в базе данных.
            Возвращает данные ранее созданных пользователей с тем же nickname-ом иои email-ом.
          schema:
            $ref: '#/components/schemas/Users'
  /user/{nickname}/profile:
    get:
      summary: Получение информации о пользователе
      tags:
        - user
      description: |
        Получение информации о пользователе форума по его имени.
      consumes: [ ]
      operationId: userGetOne
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
          schema:
            $ref: '#/components/schemas/Nickname'
      responses:
        200:
          description: |
            Информация о пользователе.
          schema:
            $ref: '#/components/schemas/User'
        404:
          description: |
            Пользователь отсутсвует в системе.
          schema:
            $ref: '#/components/schemas/Error'
    post:
      summary: Изменение данных о пользователе
      description: |
        Изменение информации в профиле пользователя.
      tags:
        - user
      operationId: userUpdate
      parameters:
        - name: nickname
          in: path
          description: Идентификатор пользователя.
          required: true
          type: string
          schema:
            $ref: '#/components/schemas/Nickname'
        - name: profile
          in: body
          description: Изменения профиля пользователя.
          required: true
          schema:
            $ref: '#/components/schemas/UserUpdate'
      responses:
        200:
          description: |
            Актуальная информация о пользователе после изменения профиля.
          schema:
            $ref: '#/components/schemas/User'
        404:
          description: |
            Пользователь отсутсвует в системе.
          schema:
            $ref: '#/components/schemas/Error'
        409:
          description: |
            Новые данные профиля пользователя конфликтуют с имеющимися пользователями.
          schema:
            $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          readOnly: true
          description: |
            Текстовое описание ошибки.
            В процессе проверки API никаких проверок на содерижимое данного описание не делается.
          example: |
            Can't find user with id #42
...
    User:
      description: |
        Информация о пользователе.
      type: object
      properties:
        nickname:
          type: string
          format: identity
          readOnly: true
          description: |
            Имя пользователя (уникальное поле).
            Данное поле допускает только латиницу, цифры и знак подчеркивания.
            Сравнение имени регистронезависимо.
          example: j.sparrow
        fullname:
          type: string
          description: Полное имя пользователя.
          example: Captain Jack Sparrow
          x-isnullable: false
        about:
          type: string
          format: text
          description: Описание пользователя.
          example: This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!
        email:
          type: string
          format: email
          description: Почтовый адрес пользователя (уникальное поле).
          example: captaina@blackpearl.sea
          x-isnullable: false
      required:
        - fullname
        - email
    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'
    UserUpdate:
      description: |
        Информация о пользователе.
      type: object
      properties:
        fullname:
          type: string
          description: Полное имя пользователя.
          example: Captain Jack Sparrow
        about:
          type: string
          format: text
          description: Описание пользователя.
          example: This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!
        email:
          type: string
          format: email
          description: Почтовый адрес пользователя (уникальное поле).
          example: captaina@blackpearl.sea
...
    Thread:
      description: |
        Ветка обсуждения на форуме.
      type: object
      properties:
        id:
          type: number
          
          description: Идентификатор ветки обсуждения.
          readOnly: true
          example: 42
        title:
          type: string
          description: Заголовок ветки обсуждения.
          example: Davy Jones cache
          x-isnullable: false
        author:
          type: string
          format: identity
          description: Пользователь, создавший данную тему.
          example: j.sparrow
          x-isnullable: false
        forum:
          type: string
          format: identity
          description: Форум, в котором расположена данная ветка обсуждения.
          readOnly: true
          example: pirate-stories
        message:
          type: string
          format: text
          description: Описание ветки обсуждения.
          example: An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?
          x-isnullable: false
        votes:
          type: number
          
          description: Кол-во голосов непосредственно за данное сообщение форума.
          readOnly: true
        slug:
          type: string
          format: identity
          description: |
            Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL).
            В данной структуре slug опционален и не может быть числом.
          pattern: ^(\d|\w|-|_)*(\w|-|_)(\d|\w|-|_)*$
          readOnly: true
          example: jones-cache
        created:
          type: string
          format: date-time
          description: Дата создания ветки на форуме.
          example: 2017-01-01T00:00:00.000Z
          x-isnullable: true
      required:
        - title
        - author
        - message
...
    ThreadUpdate:
      description: |
        Сообщение для обновления ветки обсуждения на форуме.
        Пустые параметры остаются без изменений.
      type: object
      properties:
        title:
          type: string
          description: Заголовок ветки обсуждения.
          example: Davy Jones cache
        message:
          type: string
          format: text
          description: Описание ветки обсуждения.
          example: An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?
    Post:
      description: |
        Сообщение внутри ветки обсуждения на форуме.
      type: object
      properties:
        id:
          type: number
          
          description: Идентификатор данного сообщения.
          readOnly: true
        parent:
          type: number
          
          description: |
            Идентификатор родительского сообщения (0 - корневое сообщение обсуждения).
        author:
          type: string
          format: identity
          description: Автор, написавший данное сообщение.
          example: j.sparrow
          x-isnullable: false
        message:
          type: string
          format: text
          description: Собственно сообщение форума.
          example: We should be afraid of the Kraken.
          x-isnullable: false
        isEdited:
          type: boolean
          description: Истина, если данное сообщение было изменено.
          readOnly: true
          x-isnullable: false
        forum:
          type: string
          format: identity
          description: Идентификатор форума (slug) данного сообещния.
          readOnly: true
        thread:
          type: number
          
          description: Идентификатор ветви (id) обсуждения данного сообещния.
          readOnly: true
        created:
          type: string
          format: date-time
          description: Дата создания сообщения на форуме.
          readOnly: true
          x-isnullable: true
      required:
        - author
        - message
    Posts:
      type: array
      items:
        $ref: '#/components/schemas/Post'
...
    Vote:
      type: object
      description: |
        Информация о голосовании пользователя.
      properties:
        nickname:
          type: string
          format: identity
          description: Идентификатор пользователя.
          x-isnullable: false
        voice:
          type: number
          
          description: Отданный голос.
          enum:
            - -1
            - 1
          x-isnullable: false
      required:
        - nickname
        - voice
    Slug:
      type: string
      format: identity
    Limit:
      type: number
      
      default: 100
      minimum: 1
      maximum: 10000
    Since:
      type: string
      format: date-time
    Desc:
      type: boolean
...
    Sort:
      type: string
      description: |
        Вид сортировки:

          * flat - по дате, комментарии выводятся простым списком в порядке создания;
          * tree - древовидный, комментарии выводятся отсортированные в дереве
            по N штук;
          * parent_tree - древовидные с пагинацией по родительским (parent_tree),
            на странице N родительских комментов и все комментарии прикрепленные
            к ним, в древвидном отображение.

        Подробности: https://park.mail.ru/blog/topic/view/1191/
      default: flat
      enum:
        - flat
        - tree
        - parent_tree
    Nickname:
      type: string